Only count distinct (domain+field) sniffs once
This is a proposed behavior change to only increment the sniff count once, per distinct "sniff-domain/sniffed-field".

The change alters the "inputSniffed" message handler to -update- (rather than -insert-) sniff details, if the domain+fieldName+xpath already exists.
Included is a modification to setBadge() to display the sniff count when the badge is yellow (i.e. when there are sniffs, but no leaky requests).